Fix type incompatiblity errors in memory manager
This updates the PHP memory manager based on the Native
memory manager preprocessor, and simply changes the names
of the malloc family function calls.

A MemoryManager::load function is re-added to select the
memory manager to use based on the loadtype passed.

Note: MemoryManager.pre won't currently compile for me
on the master branch because of an unknown problem with
the $extra parameter in the compile block. The MemoryManager.php
generated in this PR wasn't directly compiled, but compiled
with those lines commented out and then partially staged
with "git add -p".