Fix PHPLLVM Namespace reference in compiled code
This fixes the 2 errors that phan keeps picking up on
under CI: one call to a constructor has too many arguments,
and references to PHPLLVM inside of generated code is relative
to the current namespace. This makes the macros generate code
where the references are explicitly relative to the root namespace.