Fix PhanUnreferencedUseNormal from phan
This removes references to "use" statements that aren't referenced
anywhere in the file and fixes the PhanUnreferencedUseNormal "errors"
that are flagged. For each "use" that was removed, I first checked the
file to verify that there weren't, in fact, any references to them.

There's a bug in phan where it doesn't understand typed property
references as uses, so lines that were only referenced in that way are
simply ignored for now.